+++ src/main/java/com/backend/komeet/post/application/comment/CommentLikeService.java
@@ -1,9 +1,9 @@
-package com.backend.komeet.post.application;
+package com.backend.komeet.post.application.comment;
 
 import com.backend.komeet.infrastructure.components.RedisDistributedLock;
-import com.backend.komeet.post.model.entities.Post;
+import com.backend.komeet.post.model.entities.Comment;
 import com.backend.komeet.infrastructure.exception.CustomException;
-import com.backend.komeet.post.repositories.PostRepository;
+import com.backend.komeet.post.repositories.CommentRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
@@ -11,40 +11,40 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
-import static com.backend.komeet.infrastructure.exception.ErrorCode.*;
+import static com.backend.komeet.infrastructure.exception.ErrorCode.COMMENT_NOT_FOUND;
 
 /**
- * 게시물 좋아요 관련 서비스
+ * 대댓글 좋아요 서비스
  */
 @Slf4j
 @RequiredArgsConstructor
 @Service
-public class PostLikeService {
-    private final PostRepository postRepository;
+public class CommentLikeService {
+    private final CommentRepository commentRepository;
     private final RedisDistributedLock redisDistributedLock;
-    final String LOCK_KEY = "likePost : ";
+    final String LOCK_KEY = "likeComment : ";
 
     /**
-     * 게시물 좋아요를 처리하는 메서드
+     * 대댓글 좋아요를 처리하는 메서드
      *
-     * @param userSeq 사용자 식별자
-     * @param postSeq 게시물 식별자
+     * @param userSeq    사용자 식별자
+     * @param commentSeq 대댓글 식별자
      */
     @Async
     @Transactional(propagation = Propagation.REQUIRES_NEW)
-    public void likePost(Long userSeq, Long postSeq) {
+    public void likeComment(Long userSeq, Long commentSeq) {
         int maxRetries = 3;
         int retries = 0;
         long retryIntervalMillis = 1000; // 1초 간격으로 재시도
-        Long likeCount = null;
+        Integer likeCount = null;
 
         while (retries <= maxRetries) {
             try {
-                likeCount = processLike(userSeq, postSeq);
+                likeCount = processLike(userSeq, commentSeq);
                 if (likeCount != null) {
                     break;
                 } else {
-                    log.error("Failed to acquire lock for postSeq: {}", postSeq);
+                    log.error("Failed to acquire lock for commentSeq: {}", commentSeq);
                 }
             } catch (CustomException e) {
                 log.error(e.getMessage());
@@ -59,40 +59,43 @@ public class PostLikeService {
             }
         }
         if (likeCount == null) {
-            log.error("Failed to process like for postSeq {} after {} retries", postSeq, maxRetries);
+            log.error(
+                    "Failed to process like for commentSeq {} after {} retries",
+                    commentSeq, maxRetries
+            );
         }
     }
 
     /**
      * 좋아요 작업을 처리하는 메서드
      *
-     * @param userSeq 사용자 식별자
-     * @param postSeq 게시물 식별자
+     * @param userSeq    사용자 식별자
+     * @param commentSeq 게시물 식별자
      * @return 작업 성공 여부
      * @throws CustomException 작업 실패
      */
-    private Long processLike(Long userSeq, Long postSeq) throws CustomException {
-        Post post = getPost(postSeq);
+    private Integer processLike(Long userSeq, Long commentSeq) throws CustomException {
+        Comment comment = getComment(commentSeq);
 
         boolean lockAcquired = false;
         try {
             lockAcquired = redisDistributedLock.acquireLock(
-                    LOCK_KEY, postSeq.toString(), 10
+                    LOCK_KEY, commentSeq.toString(), 10
             );
             if (lockAcquired) {
-                if (post.getLikeUsers().remove(userSeq)) {
-                    post.setLikeCount(post.getLikeCount() - 1);
+                if (comment.getLikeUsers().remove(userSeq)) {
+                    comment.setUpVotes(comment.getUpVotes() - 1);
                 } else {
-                    post.getLikeUsers().add(userSeq);
-                    post.setLikeCount(post.getLikeCount() + 1);
+                    comment.getLikeUsers().add(userSeq);
+                    comment.setUpVotes(comment.getUpVotes() + 1);
                 }
-                return postRepository.save(post).getLikeCount();
+                return commentRepository.save(comment).getUpVotes();
             } else {
                 return null;
             }
         } finally {
             if (lockAcquired) {
-                redisDistributedLock.releaseLock(LOCK_KEY, postSeq.toString());
+                redisDistributedLock.releaseLock(LOCK_KEY, commentSeq.toString());
             }
         }
     }
@@ -103,9 +106,9 @@ public class PostLikeService {
      * @param seq 게시물 식별자
      * @return 게시물
      */
-    private Post getPost(Long seq) {
-        return postRepository
+    private Comment getComment(Long seq) {
+        return commentRepository
                 .findById(seq)
-                .orElseThrow(() -> new CustomException(POST_NOT_FOUND));
+                .orElseThrow(() -> new CustomException(COMMENT_NOT_FOUND));
     }
 }
