+++ src/main/java/com/backend/komeet/post/presentation/controller/JobBoardController.java
@@ -6,9 +6,9 @@ import com.backend.komeet.post.presentation.request.JobBoardUploadRequest;
 import com.backend.komeet.base.presentation.response.ApiResponse;
 import com.backend.komeet.post.enums.Experience;
 import com.backend.komeet.post.enums.Industry;
-import com.backend.komeet.post.application.JobBoardDetailService;
-import com.backend.komeet.post.application.JobBoardSearchService;
-import com.backend.komeet.post.application.JobBoardUploadService;
+import com.backend.komeet.post.application.jobboard.JobBoardDetailService;
+import com.backend.komeet.post.application.jobboard.JobBoardSearchService;
+import com.backend.komeet.post.application.jobboard.JobBoardUploadService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
