+++ src/main/java/com/backend/komeet/post/presentation/controller/PostController.java
@@ -8,10 +8,10 @@ import com.backend.komeet.base.presentation.response.ApiResponse;
 import com.backend.komeet.post.enums.Categories;
 import com.backend.komeet.user.enums.Countries;
 import com.backend.komeet.post.enums.SortingMethods;
-import com.backend.komeet.post.application.PostDeleteService;
-import com.backend.komeet.post.application.PostLikeService;
-import com.backend.komeet.post.application.PostUpdateService;
-import com.backend.komeet.post.application.PostUploadService;
+import com.backend.komeet.post.application.post.PostDeleteService;
+import com.backend.komeet.post.application.post.PostLikeService;
+import com.backend.komeet.post.application.post.PostUpdateService;
+import com.backend.komeet.post.application.post.PostUploadService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
