+++ src/main/java/com/backend/komeet/post/presentation/controller/ReplyController.java
@@ -3,8 +3,8 @@ package com.backend.komeet.post.presentation.controller;
 import com.backend.komeet.infrastructure.security.JwtProvider;
 import com.backend.komeet.post.presentation.request.CommentUploadRequest;
 import com.backend.komeet.base.presentation.response.ApiResponse;
-import com.backend.komeet.post.application.ReplyLikeService;
-import com.backend.komeet.post.application.ReplyUploadService;
+import com.backend.komeet.post.application.reply.ReplyLikeService;
+import com.backend.komeet.post.application.reply.ReplyUploadService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
